aiogram_app/requirements.txt
aiogram==3.1.1
aiohttp
asyncio
python-decouple